move payment helper into its own file with a namespace for braintree
