+++ app/controllers/api/stateless/braintree/subscriptions_controller.rb
@@ -5,11 +5,11 @@ module Api
         before_filter :authenticate_request!
 
         def index
-          @subscriptions = subscriptions_for_member(@current_member)
+          @subscriptions = PaymentHelper::Braintree.subscriptions_for_member(@current_member)
         end
 
         def destroy
-          @subscription = subscription_for_member(member: @current_member, id: params[:id])
+          @subscription = PaymentHelper::Braintree.subscription_for_member(member: @current_member, id: params[:id])
 
           begin
             result = ::Braintree::Subscription.cancel(@subscription.subscription_id)
@@ -23,21 +23,6 @@ module Api
             @subscription.destroy
             render json: { success: true }
           end
-
-        end
-
-        private
-
-        def customer(member)
-          ::Payment::Braintree::Customer.find_by!(member_id: member.id)
-        end
-
-        def subscriptions_for_member(member)
-          customer(member).subscriptions.order('created_at desc')
-        end
-
-        def subscription_for_member(member:, id:)
-          customer(member).subscriptions.find(id)
         end
       end
     end
