+++ app/controllers/api/stateless/payment_helper.rb
@@ -1,0 +1,25 @@
+module PaymentHelper
+  module Braintree
+    extend self
+
+    def customer(member)
+      ::Payment::Braintree::Customer.find_by!(member_id: member.id)
+    end
+
+    def payment_methods_for_member(member)
+      customer(member).payment_methods.order('created_at desc')
+    end
+
+    def payment_method_for_member(member:, id:)
+      customer(member).payment_methods.find(id)
+    end
+
+    def subscriptions_for_member(member)
+      customer(member).subscriptions.order('created_at desc')
+    end
+
+    def subscription_for_member(member:, id:)
+      customer(member).subscriptions.find(id)
+    end
+  end
+end
